Make solana-address-lookup-table-program crate bpf compatible
