+++ programs/address-lookup-table/Cargo.toml
@@ -18,9 +18,12 @@ num-traits = "0.2"
 serde = { version = "1.0.136", features = ["derive"] }
 solana-frozen-abi = { path = "../../frozen-abi", version = "=1.10.3" }
 solana-frozen-abi-macro = { path = "../../frozen-abi/macro", version = "=1.10.3" }
+solana-program = { path = "../../sdk/program", version = "=1.10.3" }
+thiserror = "1.0"
+
+[target.'cfg(not(target_arch = "bpf"))'.dependencies]
 solana-program-runtime = { path = "../../program-runtime", version = "=1.10.3" }
 solana-sdk = { path = "../../sdk", version = "=1.10.3" }
-thiserror = "1.0"
 
 [build-dependencies]
 rustc_version = "0.4"
