+++ runtime/src/bank/address_lookup_table.rs
@@ -1,10 +1,9 @@
 use {
     super::Bank,
+    solana_address_lookup_table_program::error::AddressLookupError,
     solana_sdk::{
         message::v0::{LoadedAddresses, MessageAddressTableLookup},
-        transaction::{
-            AddressLoader, AddressLookupError, Result as TransactionResult, TransactionError,
-        },
+        transaction::{AddressLoader, Result as TransactionResult, TransactionError},
     },
 };
